requirements.txt
opencv-python-headless
requests==2.24.0
Pillow
git+https://github.com/Rapptz/discord.py
PyNaCl==1.3.0
pandas
numpy
asyncio
ffmpeg